+++ CHANGELOG.md
@@ -6,10 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
-- 1 new water unit
+- 2 new water units
 	1. Troy Ounce (`t oz`)
-- 1 new coffee unit
+	2. Grain (`gr`)
+- 2 new coffee units
 	1. Troy Ounce (`t oz`)
+	2. Grain (`gr`)
 ## [1.0] - 2024-12-17
 ### Added
 - 3 new water units